requirements.txt
Django==3.0.5
newsapi-python